Mods to the alarm timestamp properties
The timestamp can be set to a predefined value when adding a new alarm
to the db.
The timestamp cannot be edited once the alarm is in the db, however, the
update operation already modifies the AlarmItem.timestamp variable with
the new value, so checking the new timestamp is trivial and under the
responsability of the developer (for syncing purposes).
An AlarmItem will have its timestamp variable as None until edited
directly.