CLI interface as a thread to launch with server
Changes to be able to launch the command line interface on its own
thread in order to be able to launch it together with the server.
This offers an additional command line argument to the application.
While originally you had to choose between running the CLI or the
server, now you can also select both.